refactor(app): extract session inspection read model
